+++ src/css/header.css
@@ -6,7 +6,7 @@
 }
 
 /* Header styles */
-.container {
+.head-cont {
   /* Styles for container */
   display: flex;
   justify-content: space-between;
