+++ src/main.js
@@ -1,4 +1,9 @@
-import { openModal, closeModal, toggleMenuHead } from './js/header.js';
+import {
+  openModal,
+  closeModal,
+  toggleMenuHead,
+  headOpenModal,
+} from './js/header.js';
 import './js/hero.js';
 import './js/abouts.js';
 import './js/benefits.js';
@@ -8,21 +13,7 @@ import './js/covers.js';
 import './js/reviews.js';
 import './js/work-together.js';
 
-(() => {
-  const refs = {
-    openModalBtn: document.querySelector('.head-btn'),
-    closeModalBtn: document.querySelector('.modal-close-btn'),
-    modal: document.querySelector('.mobile-modal'),
-  };
-
-  refs.openModalBtn.addEventListener('click', toggleModal);
-  refs.closeModalBtn.addEventListener('click', toggleModal);
-
-  function toggleModal() {
-    refs.modal.classList.toggle('is-hidden');
-    document.body.classList.toggle('no-scroll');
-  }
-})();
-
 // header menu function
 toggleMenuHead();
+// open modal function
+headOpenModal();
